makes API cmp-like
