+++ r1cs-std/src/lib.rs
@@ -61,7 +61,7 @@ pub mod pairing;
 pub mod alloc;
 pub mod eq;
 pub mod select;
-pub mod smaller_than;
+pub mod cmp;
 
 pub mod prelude {
     pub use crate::{
@@ -73,6 +73,7 @@ pub mod prelude {
         instantiated::*,
         pairing::PairingGadget,
         select::*,
+        cmp::CmpGadget,
     };
 }
